requirements.txt
keras==2.3.1
scipy==1.4.1
pandas==0.24.2
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scikit-video==1.1.11
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.2.0
imutils==0.5.3
PyYAML==5.3.1
tqdm==4.48.0
colorcet==2.0.2
google-cloud-vision==1.0.0
neural-structured-learning==1.3.1
nose2==0.9.2
